refactor: split TagManagerDialogs into individual component files
