+++ src/javascript/TagManager/TagManagerDrawer.jsx
@@ -15,7 +15,8 @@ import {
 } from '@jahia/moonstone';
 import {useTranslation} from 'react-i18next';
 import {GET_TAGGED_CONTENT, DELETE_TAG_ON_NODE, RENAME_TAG_ON_NODE} from './TagManager.gql-queries';
-import {DeleteNodeTagDialog, EditNodeTagDialog} from './TagManagerDialogs';
+import {DeleteNodeTagDialog} from './DeleteNodeTagDialog';
+import {EditNodeTagDialog} from './EditNodeTagDialog';
 import {getImpactedItemsCount} from './TagManager.utils';
 import {NodeIcon} from '~/utils/NodeIcon';
 import styles from './TagManager.scss';
